mach-exynos: cpufreq: Prevent bogus old_index caused by  policy change
For the purposes of deciding whether to completely reset the APLL,
or just change the s divisor, exynos_target determines the old
frequency index for the current frequency.

Unfortunately, as it uses the standard cpufreq table lookup code,
which does policy enforcement, this will cause a wrong index_old
if the current CPU frequency is outside of the policy limits,
for example immediately after a policy min/max frequency change.

This ensures that the current frequency is included within the
policy limits used to determine old_index, and also ensures
that the current frequency is correct by getting it from the
hardware instead of using freqs.cur.

This solves the issue on GT-I9100 and SGH-I777 kernels where
policy changes (such as SetCPU screen-off profiles) cause crashes.

This patch in its original form was written by:
Gokhan Moral <gm@alumni.bilkent.edu.tr>